refactor: refactor auth for crm integrations
